# specs/005-core-domain-crud-apis/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: AkarAI Phase 4 Core Domain CRUD API
  version: 0.1.0
  description: >
    Contract for Phase 4 core domain database and CRUD APIs. This contract excludes
    AI search, RAG, image upload/processing, OCR, email sending, dashboards,
    chatbot behavior, buyer-to-agency real-time chat, spam classification,
    lead scoring, and generated replies.
servers:
  - url: http://localhost:8000
security:
  - bearerAuth: []
paths:
...
  /agency/listings:
    get:
      summary: List tenant listings
      tags: [Agency Listings]
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
      responses:
        "200": { description: Paginated tenant listings }
    post:
      summary: Create listing
      tags: [Agency Listings]
      responses:
        "201": { description: Listing created }
        "403": { description: Support employees cannot create listings }
  /agency/listings/{listing_id}:
    get:
      summary: Get tenant listing
      tags: [Agency Listings]
      parameters:
        - $ref: "#/components/parameters/ListingId"
      responses:
        "200": { description: Listing }
    patch:
      summary: Update tenant listing
      tags: [Agency Listings]
      parameters:
        - $ref: "#/components/parameters/ListingId"
      responses:
        "200": { description: Listing updated }
    delete:
      summary: Archive tenant listing
      tags: [Agency Listings]
      parameters:
        - $ref: "#/components/parameters/ListingId"
      responses:
        "204": { description: Listing archived }
  /agency/listings/{listing_id}/photos:
    get:
      summary: List listing photo metadata
      tags: [Agency Listings]
      parameters:
        - $ref: "#/components/parameters/ListingId"
      responses:
        "200": { description: Listing photo metadata }
    post:
      summary: Create listing photo metadata
      tags: [Agency Listings]
      parameters:
        - $ref: "#/components/parameters/ListingId"
      responses:
        "201": { description: Photo metadata created }
...
  /listings:
    get:
      summary: Public active listing search
      tags: [Public Listings]
      parameters:
        - name: location
          in: query
          schema: { type: string }
        - name: min_price
          in: query
          schema: { type: number }
        - name: max_price
          in: query
          schema: { type: number }
        - name: bedrooms
          in: query
          schema: { type: integer }
        - name: bathrooms
          in: query
          schema: { type: integer }
        - name: property_type
          in: query
          schema: { type: string }
        - name: listing_purpose
          in: query
          schema: { type: string }
        - name: furnishing
          in: query
          schema: { type: string }
        - name: min_area_size
          in: query
          schema: { type: number }
        - name: max_area_size
          in: query
          schema: { type: number }
        - name: sort
          in: query
          schema:
            type: string
            enum: [newest, price_asc, price_desc, area_size_asc, area_size_desc]
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
      responses:
        "200": { description: Paginated active listing results }
        "429": { description: Search rate limit exceeded }
  /listings/{listing_id}:
    get:
      summary: Public listing detail
      tags: [Public Listings]
      parameters:
        - $ref: "#/components/parameters/ListingId"
      responses:
        "200": { description: Active listing detail }
        "404": { description: Listing missing or not active }
  /me/saved-listings:
    get:
      summary: List my saved listings
      tags: [Saved Listings]
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
      responses:
        "200": { description: Paginated saved listings }
  /me/saved-listings/{listing_id}:
    put:
      summary: Save listing
      tags: [Saved Listings]
      parameters:
        - $ref: "#/components/parameters/ListingId"
      responses:
        "200": { description: Listing saved }
    delete:
      summary: Unsave listing
      tags: [Saved Listings]
      parameters:
        - $ref: "#/components/parameters/ListingId"
      responses:
        "204": { description: Listing unsaved }
  /me/comparison-sessions:
    get:
      summary: List my comparison sessions
      tags: [Comparisons]
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
      responses:
        "200": { description: Paginated comparison sessions }
    post:
      summary: Create comparison session
      tags: [Comparisons]
      responses:
        "201": { description: Comparison session created }
  /me/comparison-sessions/{session_id}:
    get:
      summary: Get comparison session
      tags: [Comparisons]
      parameters:
        - $ref: "#/components/parameters/SessionId"
      responses:
        "200": { description: Comparison session }
    patch:
      summary: Update comparison session
      tags: [Comparisons]
      parameters:
        - $ref: "#/components/parameters/SessionId"
      responses:
        "200": { description: Comparison session updated }
    delete:
      summary: Delete comparison session
      tags: [Comparisons]
      parameters:
        - $ref: "#/components/parameters/SessionId"
      responses:
        "204": { description: Comparison session deleted }
  /me/comparison-sessions/{session_id}/items:
    post:
      summary: Add comparison item
      tags: [Comparisons]
      parameters:
        - $ref: "#/components/parameters/SessionId"
      responses:
        "201": { description: Comparison item added }
        "409": { description: Comparison session already has four items }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Page:
      name: page
      in: query
      schema: { type: integer, minimum: 1, default: 1 }
    PageSize:
      name: page_size
      in: query
      schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
    ListingId:
      name: listing_id
      in: path
      required: true
      schema: { type: string, format: uuid }
...
    SessionId:
      name: session_id
      in: path
      required: true
      schema: { type: string, format: uuid }
...